refactor:Cleanup unsed function and components.
