+++ src/components/auth/index.ts
@@ -1,2 +1,1 @@
 export * from "./AcceptInviteForm";
-export { default as SignupForm } from "./SignupForm";
